atomisp2: cleaning up the atomisp code during atomisp_release
BZ: 79817

Currently sbudev init with val = 0 is called just before the
sensor power off. But the sensor drivers are doing all the
un-initialisation routines as part of power off and there is
no need to call the init at that point. So this patch removes the
subdev init with val = 0 call being called during atomisp_release

Change-Id: I1b12ae59b64173767845bc3581c88996938dd99e
Orig-Change-Id: I52f9e244c66a0ea4e33c9835fee2182d8123bc6c
Signed-off-by: Vinod Govindapillai <vinod.govindapillai@intel.com>
Acked-by: Tuukka Toivonen <tuukka.toivonen@intel.com>
Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Reviewed-on: http://android.intel.com:8080/91854
Reviewed-by: Kaartinen, Jukka O <jukka.o.kaartinen@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Patale, Pavan <pavan.patale@intel.com>
Tested-by: Patale, Pavan <pavan.patale@intel.com>
Reviewed-by: Koskinen, Ilkka <ilkka.koskinen@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>