mt9m114: increased frame skip count for 1024x576 resolution
BZ: 113437

When the preview starts there is a period when the AE is not yet
converged, which causes a user-noticeable dark flashing in the
preview, if the lighting is no sufficient. This happens e.g. indoors.
To address the issue, the frame skipping is increased from 1 frame
to 3 frames, so that the AE has time to converge.

Porting from: http://android.intel.com:8080/#/c/107357/1

Change-Id: I4d841368ee547db899c77f8d024b566333538d2e
Orig-Change-Id: I2671c0a2e77b742ebee87902c5ab7dd3444422b4
Signed-off-by: Bin Han <bin.b.han@intel.com>
Reviewed-on: http://android.intel.com:8080/112546
Reviewed-by: Vtminen, LassiX <lassix.vtminen@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>