ov8835: Select the lowest FPS setting for MERR A0
BZ: 117595

ISP timeout out happened on PRH when using a higher FPS setting like 8M@25FPS
and 6M@24FPS. This is mainly because ISP1.5 firmware for ISP2400A0 is not
enough for these settings and ISP1.5 will not be used in the future ISP2400
projects. So add this workaround to make lowest FPS setting can be always
chooesn for ISP2400A0.

And this patch addes a 19FPS setting for 6M.

Change-Id: I68de1103835bdcf3a1bf5e4788607f88cbf862c4
Signed-off-by: Bin Han <bin.b.han@intel.com>
Reviewed-on: http://android.intel.com:8080/114600
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Kaartinen, Jukka O <jukka.o.kaartinen@intel.com>
Reviewed-by: Govindapillai, Vinod <vinod.govindapillai@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>