atomisp2: clean up checks for 2ppc and raw-binning
BZ: 120773

The code to check whether to enable CSS 2ppc and raw-binning modes, is
inconsistent with regards to input and output resolutions. Some of the
checks use non-padded output resolutions, while some checks assume a 16
pixel padding.

Add a clarification that the resolution used for the checks is the
sensor input resolution, and that amount of padding should not be
assumed.

As a functional change, this patch disables 2ppc mode for 8MP output
resolution on ISP2300, as this combination produces invalid output.

Change-Id: Ic3b6e2f3cd78110e102c739a8ec5bcc850280b61
Signed-off-by: Kai Vehmanen <kai.vehmanen@intel.com>
Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Reviewed-on: http://android.intel.com:8080/116748
Reviewed-by: Ailus, Sakari <sakari.ailus@intel.com>
Reviewed-by: Koskinen, Ilkka <ilkka.koskinen@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>