atomisp2: implement multiple atomisp_sub_device instances.
BZ: 111659

This is the preliminary patch for atomisp driver to support
dual(multiple) streams.

Implement multiple atomisp_sub_device instances to represent
different streams.

The number of atomisp_sub_devices is decided in runtime by macro
IS_ISP2400(); If it is isp2400, the number is 2, or it is 1 for
isp2300.

Change-Id: I8c682ad5fd37b2d99da18c1a65f0ec09ea9a290e
Signed-off-by: Tao, Jing <jing.tao@intel.com>
Signed-off-by: Antti Laakso <antti.laakso@intel.com>
Reviewed-on: http://android.intel.com:8080/122852
Reviewed-by: Jiang, ZiyuX <ziyux.jiang@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>