[PORT FROM R42B] atomisp2: disable video dis due to sensor's limitation
BZ: 121572

We need to disable dis flow when sensor can not output enough
envelop.

The issue is found on BTY's front sensor whose max output
is 1932x1092. So for 1080p, we need to disable dis in driver to prevent
the dis calculation flow to output final effect resolution to be 1600x900.

For CSS2.0, the digital zoom/DVS2 lib uses the original dis flow, though dvs
is disable due to sensor limitation, digital zoom may be enabled.

Depends on patches: 132134 129805

Change-Id: If0ca628255c3f13e7fdbde84c7c8bc6740292fb2
Orig-Change-Id: Iba2d8327a50bf68cc6f08c903e1e02f3f996feb6
Signed-off-by: Tao, Jing <jing.tao@intel.com>
Signed-off-by: Shuguang Gong <shuguang.gong@intel.com>
Reviewed-on: http://android.intel.com:8080/132136
Reviewed-by: Kuppuswamy, Sathyanarayanan <sathyanarayanan.kuppuswamy@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>