atomisp2: Add initialization value for loop variable i
BZ: 122890

The local variable "i" is not initialized in atomisp_acc_load_extensions().
It may be referenced with undefined value in the error handling part if link list isp->acc.fw is empty
Give "i" a default value of -1 so that the error hanling code can be bypassed if no list items are processed.

Change-Id: Ife4037e973de3c91c7fa3eea1c9bec9d916d41f0
Signed-off-by: Tianshu Qiu <tian.shu.qiu@intel.com>
Reviewed-on: http://android.intel.com:8080/131036
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: Kuppuswamy, Sathyanarayanan <sathyanarayanan.kuppuswamy@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>