atomisp2: fix ION compile issues
BZ: 136879

ION inteface has changed a lot since JellyBean, some APIs have been
changed/renamed. This patch checks all the compile issues and aligns to the
latest ION interface defined in the inlcude/linux/ion.h.

One last compile issus not fixed is that we do not have a ion_device defined
by ION driver which is not implemented yet. This is expected to be fixed when
integrating the ION buffer sharing with GFX/Video team.

Change-Id: I83b9c9c07316d2bba12c6f542283fb8dd5b71f1a
Signed-off-by: Bin Han <bin.b.han@intel.com>
Signed-off-by: Antti Laakso <antti.laakso@intel.com>
Reviewed-on: http://android.intel.com:8080/131620
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>