imx132: Add 1636x1096 setting for 480p video mode.
BZ: 146101

To keep FOV consistent, HAL selects 1620x1080 for IMX132 480p video mode.
The VF output is set to 720x480 which is finally used for video recording.
Currently there is no suitable setting for this case.
Digital zoom will get incorrect calculation result if unsuitable setting is selected.
So add a new setting for IMX132 which meets this case.

Change-Id: I9f2cb6d98e3df6293630c1eccd0934265b9c1e59
Signed-off-by: Tianshu Qiu <tian.shu.qiu@intel.com>
Acked-by: Vinod Govindapillai <vinod.govindapillai@intel.com>
Signed-off-by: Antti Laakso <antti.laakso@intel.com>