[PORT FROM R43S]atomisp2: Workaround to prevent system crash in ISP timeout
BZ: 123401

Sometimes camera subsystem goes to odd state in pipe shutdown.
From driver point of view there is a timeout at stopping
phase. Most of the times ISP status and registers
can still be accessed but sometimes any access to
ISP registers causes IA core to hang. This leads
to reboot.
To avoid system wide reboot all the accesses to camera
subsystem are avoided in case of timeout.
The only thing what we can do is to drive the camera
sub system to d0i3 state via ospm command.
This is a workaround since the reason for ISP crash is
not known.

Change-Id: Ib2e8e8970c29cfdf8bf3b87516f3d4a31bf6a723
Orig-Change-Id: I86387d75777cf4e5ae106b3a12924cd4646fac1e
Signed-off-by: Samu Onkalo <samu.onkalo@intel.com>
Signed-off-by: Antti Laakso <antti.laakso@intel.com>
Acked-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Acked-by: Ilkka Koskinen <ilkka.koskinen@intel.com>