atomisp2: Query camera capabilities during sub device probe.
BZ: 158552

This patch adds querying for the camera capabilities of each
camera sub device. The sensor driver needs to provide a pointer
to the capability structure. The pointer is saved at
struct atomisp_input_subdev.camera_caps.
This patch also adds a new field in struct atomisp_sub_device
named sensor_curr which indicates which sensor is selected by
the user. The default value is 0. It should be set in vidioc_s_input
call if multiple sensors are connected to the same port and
user wants to select the ones other than the first.

Change-Id: I00971b5418156ea5cbff649e893eccb545efe1c8
Signed-off-by: Tianshu Qiu <tian.shu.qiu@intel.com>