atomisp2: release empty MMU L2 page tables when camera application exits
BZ: 161621

For CSS2.X, the ISP FW is loaded into memory and MMU mapping is setup when
atomisp kernel driver is insmoded. So, when releasing all the atomisp file
handles, driver can not do the full cleanup of MMU which means some empty MMU
L2 page tables can not be released, so that this is kind of memory leakage.

This patch adds the hmm_cleanup_mmu_l2() function to provide the releasing of
empty MMU L2 page table and it will be called when all atomisp driver file
handles get released. And it is used only by CSS2.X because for CSS1.5
hrt_isp_css_mm_clear() will do the full cleanup of MMU.

Change-Id: I5c025b8bbf1b951995dbafa36c8555195a0a33fb
Signed-off-by: Bin Han <bin.b.han@intel.com>