imx: fix NULL pointer exception
BZ: 164654

The driver checked whether the function pointer
dev->platform_data->msr_file_name was NULL or not, but it did not
check whether the value returned by the function was NULL.
load_msr_list() was generating NULL pointer dereference if the
NULL pointer was passed to it. Fix this.

Change-Id: I56a89c3bf64e414f32b951f6f8a8716a004e9754
Signed-off-by: Tuukka Toivonen <tuukka.toivonen@intel.com>
Reviewed-on: https://android.intel.com/156220
Reviewed-by: Vtminen, LassiX <lassix.vtminen@intel.com>
Reviewed-by: Laakso, Antti <antti.laakso@intel.com>
Reviewed-by: Govindapillai, Vinod <vinod.govindapillai@intel.com>
Reviewed-by: Kaartinen, Jukka O <jukka.o.kaartinen@intel.com>
Reviewed-by: Lampila, KalleX <kallex.lampila@intel.com>
Tested-by: Bourahmani, Karim <karim.bourahmani@intel.com>