atomisp2: add support of 0x117a TNG IUNIT device ID
BZ: 164992

0x117a IUNIT device is also TNG IUNIT which has different ISP freq setting,
the max freq at Vmin is 200MHz and the max freq is 400MHz. This patch adds the
driver support of this IUNIT device as well as the dfs_table for it.

And to align with other MERR IUNIT device name,  the
ATOMISP_PCI_DEVICE_SOC_MRFLD_FREQ_LIMITED is changed to
ATOMISP_PCI_DEVICE_SOC_MRFLD_1179.

Change-Id: I41c0ac133ecf1446ca6c5c9f6efa8943d70015af
Signed-off-by: Bin Han <bin.b.han@intel.com>