atomisp-ng: mmu: Serialise write access to the page tables
The page tables require serialising write access to them. Read access does
not need to be serialised since what ends up to the page tables (L1 and L2)
is never released (or until module unload time).

The memory allocation for L2 page tables is not serialised. If the entry in
the L2 page tables does exist, the allocated (extra) page table is released
and the existing one is used.

Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Acked-by: Jukka Kaartinen <jukka.o.kaartinen@intel.com>