atomisp2: fix check for ATOMISP_HW_REVISION
BZ: 169851

ATOMISP_HW_REVISION is in upper bits of the field media_dev.hw_revision.
The lower bits contain stepping. In a few places, the constant,
against which the comparison is made, is shifted properly but the
low bits of hw_revision are not masked, resulting comparison failure if
the stepping is nonzero.  Fix this by introducing a new macro.

Change-Id: I223129b130627cc8a2c156b4406076afa35a7fce
Signed-off-by: Tuukka Toivonen <tuukka.toivonen@intel.com>