atomisp2: fix EOF event handling
BZ: 176995

When atomisp2 driver gets end of frame (EOF) event from camera
subsystem (CSS), the CSS event does not provide the corresponding
pipe number and the subdevice (asd) has to be searched based on
CSS port id instead. __get_asd_from_port() does this.

However, __get_asd_from_port() was comparing the CSS port id to
v4l2 input set by VIDIOC_S_INPUT which is a different thing. Due
to this the correct asd was not found, EOF event was not sent,
and 3A was not working in user space because it didn't receive the
events.

This patch:
- Fixes the above by getting the CSS port id for comparisons
- Fixes the incorrect comment about input_curr (it is not MIPI port id)
- Cleans up the EOF event handling similarly to the handling of other
  events
- Replaces MAX_STREAM_NUM with num_of_streams (the prior is incorrect)

Change-Id: I711ab9f7b37aaf31d0e1995427dc2f213fb239bc
Signed-off-by: Tuukka Toivonen <tuukka.toivonen@intel.com>