atomisp2: extend hmm_vmap to support cached mapping
BZ: 177697

There is an requirement that atomisp driver can do cached mapping from ISP
virtual address to IA virtual address in order to accelerate IA performance on
accessing to the buffer, so this patch adds the cached flag parameter for the
functions and using the vmap to do cached mapping accordingly.

And there is a new API hmm_flush_vmap is added to flush/invalidate the cached
memory.

Change-Id: Ibc4dd4aef4a66119f29e2b5b0f2a9a1ce9adfd7f
Signed-off-by: Bin Han <bin.b.han@intel.com>