atomisp2: Revise video pipe remapping logic.
BZ: 174143

This patch revises the remapping logic for video
pipe in __atomisp_get_pipe.
stream_id should only be used to map video pipes
when the sensor output stream number larger than 1.
If the stream number is equal to 1,
the cases that copy pipe is used need to be taken
into consideration.

Change-Id: I23a8b90b1e424efdfca0e932b40b856900bbbbca
Signed-off-by: Tianshu Qiu <tian.shu.qiu@intel.com>