atomisp2: fix the coding issue when checking different MERR SKUs
BZ: 179717

When checking the device id of different MERR SKUs, the device id should
not do the AND-opertion with ATOMISP_PCI_DEVICE_SOC_MASK, beacuse different
MERR SKUs diff by the low 3 bits of device id.

And this patch removed the support of ISP2400A0 when initiate the hw_revision,
because ISP2400A0 is not supported any more.

Change-Id: I742199ec332e9ddb3ce6ea1ffed378824edf77d9
Signed-off-by: Bin Han <bin.b.han@intel.com>