atomisp2: Fix dual online still capture css panic
BZ: 180452

It is found that in stress test of dual stream still capture, css will get
panic ramdonly.

The reason is that in the stage of ia_css_uninit, some allocated css parameters
have not been freed, which causes css assert panic in refcount check.

The Root cause is:
1: For stream 1, it is at the stage that after set_fmt but before stream on is called;
2: For stream 2, it is calling stream off while the atomisp_reset() will get called.

For stream 1, the css stream is created at this time, but at the time atomisp_reset() is called,
which will call ia_css_uninit, there is no way to destroy the created css stream for stream 1, as
the css stream destroy is called in stream off.

Fix is to add the variable stream_prepared to record whether the css stream is created,
1: at the time atomisp_reset is called, it will check if other stream has css stream prepared, will
destroy the streams.
2: at the time atomisp_css_start is called, it will check if stream_prepared, if not, will recreate the
streams.

Change-Id: Ie183edfe9daf15ed08889978a0375cd2c70acceb
Signed-off-by: Tao, Jing <jing.tao@intel.com>