atomisp2: remove the 3A and DIS statistics translation from ISR to IOTCL routine
BZ: 177700

ISR is usually timing critical, and driver needs to make sure it does the
least. In the past, atomisp driver is doing the 3A and DIS statistics
translation in the ISR thread (bottom part of the ISR), but this is not actually
necessary, because driver can save the statistics buffer and do the
translation when userspace requires the statistics results by IOCTL.

This patch creates two link list for each type statistics buffers, one for
free buffers in which the latest statistics data buffer will be in the head of
list, the other one is for the buffers which are enqueued into CSS. ISR thread
always take the latest done statistics buffer from the CSS link list and put
it to the head of the free link list for userspace IOCTL. and takes the tail
of the free list and enqueued to CSS. because the overall statistics buffer
number is 1 + CSS_Q_DEPTH, so there will be always one free statistics buffer
which contains the latest statistics in the free link.

Change-Id: Ia67f6b40538e4205cafa7663ec3941711b26243c
Signed-off-by: Bin Han <bin.b.han@intel.com>