atomisp2: enable DVS for SDV use case
BZ: 182762

Several changes are included to enable DVS for SDV:

1. The DVS envelope is actually cropped from BDS output. Therefore, no need
to set the dvs_envelope(in pipe configuration) to 20% output resolutions,
because it is used to crop from sensor input. As requirement, set the
dvs_envelope to 12x12 and calculate the real DVS envelope in atomisp_param()
and atomisp_get_dvs2_bq_resolutions() which will be used to configure DVS2
library by camera HAL.

2. When choosing BDS downscaling factor for DVS on case, take at least 10%
output resolution as envelope into consideration. Previously, the minimum
is 18%, but it actually does not work for 1080P video with DVS on.

3. Remove the 1.25X BDS which is actually not implemented by FW. Otherwise,
assertion failed with 1080P DVS case.

Change-Id: Ie2d301006ebc906c5c97a19ecc1bdb275d74f4c8
Signed-off-by: Bin Han <bin.b.han@intel.com>