m10mo: Fix checksum calculation
BZ: 182736

The same workaround as we have in fw-start is needed
also in check sum calculation. With the WA in place
a proper FW image returns checksum as 0.

FW checksum is now checked at probe function. Currently it is just
a warning message if the checksum is not zero. Automatic update
of the corrupted or missing FW is not yet implemented.

Change-Id: I1cf18e903135d1b808a742a7400a2cc4b53dd0d7
Signed-off-by: Samu Onkalo <samu.onkalo@intel.com>