atomisp2: send the exp_id of EOF event to userspace
BZ: 183204

Camera HAL needs the exp_id of EOF event to align with the 3A statistics and
frames, so send it out to camera HAL by event.u.frame_sync.frame_sequence.

Previously the frame_sequence is used to record a linear sequence number of
eof event(by the eof_count), but considering camera HAL does not need it and
the sequence of v4l2_event is doing the same thing, so change it to the exp_id
and remove the eof_count.

Change-Id: Ica5571d86e731c1c1d7894e08847e6e56ff7a4e3
Signed-off-by: Bin Han <bin.b.han@intel.com>