atomisp2: utilize vmap to optimize the performance of hmm load/store/set
BZ: 182986

Previously, we use the kmap to do un-cached memory mapping by page and then do
memcpy at page basis, this is not optimized as un-cached memcpy is slow and
does not check if the buffer has already been mapped by vmap.

This patch first check if the buffer has been mapped by vmap, if so, do memcpy
directly use the vmap_addr. If not, try to do a cached vmap and then do the
memcpy and cache flush at last. But if for some reason the vmap failed,
fallback to the old kmap way.

Usually, the isp virtual address passed to hmm_load and hmm_store is not the
start address of whole memory block allocated through hmm_alloc, so in this
case, current hmm_vmap() does not work with such memory address because hmm_vmap
only compares the start address of the whole memory block. So this patch
extends hmm_vmap/hmm_flush_vmap/hmm_vunmap to support search if the address is
in the range of an allocated buffer object and do the vmap/vunmap/cache_flush
correspondingly.

Change-Id: Ica1162eb3f4ec65cdf3f7bac5c0c738321959f6c
Signed-off-by: Bin Han <bin.b.han@intel.com>