atomisp2: fix incorrect capture_pad for offline video mode
BZ: 186461

Previously, we thought camera HAL will always configure the capture pipe under
SDV mode, but this is not true. If camera HAL enables the offline video mode
without configuring the capture pipe, the capture pipe resolutions and format
will be invalid and it should not be taken as the main capture pad. So, this
patch checks if the capture pad resolutions are configured properly, then it
will be taken as the main capture pad, otherwise, video pipe will be taken as
the main capture pad.

Change-Id: Ief8f015b08127c60f8e3f7dd4a67b86b680499c0
Signed-off-by: Bin Han <bin.b.han@intel.com>