atomisp2: group CSS parameters into together
BZ: 177834

This patch intends to group all CSS related parameters into a signle
parameter, this is because we are having a per_frame setting feature which
needs atomisp2 driver has the capbility to cache all the parameters which need
to apply on specific frames. To group all CSS related parameters into one
signle struture helps to make this task easier.

As well, this patch separates the copy userspace parameter to internal
parameter buffer and apply this parameter buffer into CSS configuration, this
is also for per_frame setting feature.

BTW: atomisp_shading_correction() is not used at all, so this patch removes
it.

Change-Id: Ifab5716686378a857c3961b59f60eee7fba88522
Signed-off-by: Bin Han <bin.b.han@intel.com>