atomisp2: add support of per_frame_setting mode
BZ: 177834

This patch includes following changes for per_frame_setting mode support:

1  Add a new V4L2_CID_PER_FRAME_SETTING for camera HAL to enable or disable
the per_frame_setting mode. In this mode, each parameter shall have a frame
pointer to point which output frame to apply this parameter.

2  Apply the isp_config_id to CSS parameter and send the isp_config_id of each
frame to camera HAL by reserved2 field of the v4l2_buffer structure which is
used by camera HAL to compare it with the isp_config_id in previous
parameter.

3  In atomisp_qbuf, check if per_frame_setting enabled and current buffer is
main buffer of the pipe, then put the buffer into waiting list if ISP is not
streaming on and there is no valid parameters for it. Because CSS/FW needs
"set per_frame parameter" before "enqueuing the frame".

4  Each time in atomisp_streamon with per_frame_setting mode, check if there
is buffer and parameter matching and then set parameters to CSS and enqueue
buffers to CSS.

5  In atomisp_set_parameters, check if per_frame_setting enabled, cache all
the parameters and put it to the waiting list until ISP is streaming on and the
specific frame is enqueued to driver. After the parameter is set to CSS, this
function will also enqueuing the buffers to CSS.

6  When ISP stops streaming or the camera HAL disables the per_frame_setting
mode, driver will flush the parameter waiting list to clean up all the not
applied parameters. Also, when atomisp_streamoff, clean up the buffer waiting
list just like activeq.

7  Do not update parameter in atomisp_isr_thread, as per_frame_setting
parameter needs to set the CSS before the corresponding frame buffer is
enqueued into CSS.

Change-Id: I755fcebdc22cabcb92d085f6b75f2a6ae71a10a3
Signed-off-by: Bin Han <bin.b.han@intel.com>