m10mo: Add v4l2 ctrl ID for ZSL capture
BZ: 187791

The ZSL capture in M10MO is different. It can output jpeg encoded
image. But this is different from the current supported capture
methods in atomisp. So we add av4l2 ctrl to pass the capture
information to m10mo which in turn will output jpeg image. This
method works only for the ZSL capture.

Change-Id: Ie605ad6946b9ea4693899189492171d61c52f508
Signed-off-by: Vinod Govindapillai <vinod.govindapillai@intel.com>