[PORT FROM R44C] m10mo: [HACK] return max supported size as the VC1 width and size
BZ: 195418

For HDR and LLS modes, m10mo sends yuv422 frames through VC1 when
switched to the respective modes. Currently we are configuring
the stream in CSS for jpeg capture with fixed size and height. So
we will not able to reconfigure the CSS without resetting it. Till
we figure out an efficient mechanism m10mo will allocate maxmum
buffer size even for jpeg capture which 13MP * 2 for yuv422.

Change-Id: I4a544af7db451e0b019e848e1984537b87a49420
Signed-off-by: Vinod Govindapillai <vinod.govindapillai@intel.com>