ov680: update the additional settings with 0506 fw.
BZ: 192991

Update the additional setting settings based on 20140506 fw.
Default setting will enable the embedded line.
But, since the embedded line still doesn't work well on new
input system, another setting to turn off the embedded line
is used, so that it will not block the 680 work.

Optimized the fw loading process and the cost time is reduced.

Change-Id: Ibc3b5bf226b3390f032387ccdedbc4aa1adf1715
Signed-off-by: Lin Xie <lin.xie@intel.com>