[PORT FROM R44C] m10mo: handle unsupported resolution by returning an error.
BZ: 195821

Currently with m10mo atomisp is running in copy mode. So it cannot
do any post processing. So far when a specific resolution is asked
from m10mo, the nearest matching resolution is sent assuming that
the isp can downscale it. This patch modifies that behavior by
returning an error in case if any unsupported resolution is
queried from m10mo.

Change-Id: I1f4cdf33a63741b54c7eed1521b61f64f66d3193
Signed-off-by: Vinod Govindapillai <vinod.govindapillai@intel.com>