atomisp2: Remove ISP maximum resolution limitation
BZ: 192186

This patch removes the limitation for ISP maximum
resolution. The current maximum definition is 13M
which is out of date. For some projects we will
support larger resolutions. Since CSS API does not
specify there is a upper limitation of the resolution,
we should not limit it in atomisp driver either.

Change-Id: I1171c91268f39ee297fa36e4220ec3bdd80ee184
Signed-off-by: Tianshu Qiu <tian.shu.qiu@intel.com>