atomisp: stream on sequence adjustment for depth camera
BZ: 201899

For Kevlar project, there are 2 sensors ov8858 and ov680.
There is a HW sync link between them. ov680 will output frame when
sync signal from ov8858 arrives. Also there will be a frame count
in sensor embedded data using to sync frames in HAL layer.
so if ov8858 is streamed on before ov680, the frame count will not match any
more.

The solution is create a ioctl for HAL to notify driver that it will
be in depth mode.
In this mode, atomisp will skip sensor stream on at the 1st asd(stream) stream-on. When 2
asds(streams) are streamed on, stream on ov680 first and then ov8858.

Change-Id: I3413dce8c144c8cc11792d5b730720fbafb9cefa
Signed-off-by: Jianxu Zheng <jian.xu.zheng@intel.com>