atomisp2: do not stop WDT in atomisp_release()
BZ: 199677

In ZSL mode, camera HAL v3 may release the video0 and video1 before streaming
off the video2, which means it closes capture video nodes but keeps preview is
streaming. At that time, WDT gets stopped and this is not expected.

atomisp_streamoff() will choose the proper time to stop WDT, so we do not need
and should not stop WDT in atomisp_release().

Change-Id: I65fb6ad251af144f165c963f3f4a9c844a466ec7
Signed-off-by: Bin Han <bin.b.han@intel.com>