ov8858: Update CSI clock speed and fix gain clamping
BZ: 201123

The CSI clock frequency needs to be changed to 420 MHz in order to
comply with EMC requirements and the registers controlling the analog
functionality of the sensor has been updated to match the clock
change.

Arrange the register addresses in ascending order.

Additionally, fix the clamping value for the analog gain so we can
use the full dynamic range.

Change-Id: I42b914927ce8287a78b1c678b64b29dcb3311a47
Signed-off-by: Soren Friis <soren.friis@intel.com>