[PORT FROM R44C] atomisp2: Proper handling of the sensor start failure
BZ: 203130

If the sensor start fails we can't call atomisp_reset.
It is a low level reset for the isp and the control layer
doesn't expect low level reset at that point.

Instead, set status to not streaming and return an error code.
ISP will be then closed in controlled manner.

Change-Id: I4ff997885f9e50e5cb90ed2dfd20b788fa0a7166
Signed-off-by: Samu Onkalo <samu.onkalo@intel.com>