atomisp2: remove guards on isp_has_started()
BZ: 206073

This patch removes the check on isp_has_started() in the stop
function. There is no need to check for isp_has_started() other than
in the case where the input FIFO is used for file injection.

And this patch also removes the guards on sh_css_hrt_system_is_idle() which
actually just a waste of time before powering down the ISP.

Change-Id: I1d7f5ebe766f6204806ba9316f7593ace20301f7
Signed-off-by: Bin Han <bin.b.han@intel.com>