m10mo: initial flash mode setting in monitor mode
BZ: 196788

The flash mode is not properly set initially when going to monitor
mode. The first flash mode might get set to the driver before the
preview is started, so we must delay setting the flash mode initially,
until monitor mode is running. This is similar to the AF mode setting.

Change-Id: Ib2fafecf787c7a9fb0413065658d2b66278fe5d2
Signed-off-by: Lassi Väätämöinen <lassi.i.vaatamoinen@intel.com>