atomisp2: meta data variable is not re-initialized correctly
BZ: 208900

The condition for 3A/meta data buffer allocation is
if 3A needs re-allocation or if meta size is different with previous one.

For SoC sensor there is no 3A buffer allocation,
also the meta data size is not re-initialized to 0
in atomisp_release(). it would cause the meta data
only works for the 1st time.

Change-Id: Ibb30f6dd4fc31991b0897dad3109415df2a17c73
Signed-off-by: Jianxu Zheng <jian.xu.zheng@intel.com>