atomisp2: Support for buffer allocating and cache flag for ACC memory mapping
BZ: 208675

This patch adds support for buffer allocating and cache
flag in ACC memory mapping. If user pointer is not assigned,
a private buffer is allocated and mapped to CSS. User
needs to call mmap passing css_ptr as the offset to map the
buffer to user space. If ATOMISP_MAP_FLAG_CACHEABLE is set
in the flags, the buffer is mapped as cacheable. Otherwise,
it is mapped as non-cacheable.

Change-Id: I91feab099f6dc84cb2544fc0ad8b397e0566b0af
Signed-off-by: Tianshu Qiu <tian.shu.qiu@intel.com>