[PORT FROM R44B] atomisp2: limit maxinum ISP freq to 320M and try more times when adjusting frequency fails
BZ: 220426

With TI PMIC on CR2.2, ISP DFS can fails when adjusting the ISP frequency to 400MHz.
This could be caused by not enough voltage on turbo mode. Limiting ISP highest
frequnency to 320MHz can avoid this issue. And, to improve the stability, DFS can be tried
more times when failing.

Change-Id: Ia26a2d416bd64b2a3a7e95add349019aede7e4d4
Signed-off-by: Shuguang Gong <shuguang.gong@intel.com>
Reviewed-on: https://android.intel.com:443/235889