atomisp2: [WORKAROUND] do not hang the system if exp_id from FW is invalid
BZ: 209274

There is a FW issue reported BZ 3300, which is in SAND project, sometimes when
testing testGcam, there will be a frame whose exp_id is invalid (a quite big
value larger than 250), driver should not hang the system when this happens.
Therefore, changing BUG_ON to WARN_ON as a workaround before FW can fix the
issue.

Change-Id: Ibcddbbffc49699cd91912e5f6b9d8d9fb1a63002
Signed-off-by: Bin Han <bin.b.han@intel.com>