atomisp2: ann: fix stepping information in hw_revision
BZ: 209387

Anniedale (Moorefield) stepping was always set to A0 which is incorrect
for newer devices. Check the revision and set to B0 for B0 and later
devices.
TODO: AX1 and K0 revisions are still not properly handled, also not other
SoCs such as CHV, but this is sufficient for now.

Change-Id: I59a3bd0bfa3c2dc9aea41cd2a638394b18287133
Signed-off-by: Tuukka Toivonen <tuukka.toivonen@intel.com>