dw9718: Change modes to reduce noise and improve performance
BZ: 211899

The switching mode used in the VCM controller is causing
interference in many of the radio frequency bands. Change it to
linear mode instead, which removes the noise.

Change also the control mode to SAC4 instead of Direct mode in order
to reduce the amount of ringing movement of the lens during focus
operations.

Change-Id: I9235dc119c0882bb50ebf52ca8402e8769fa5590
Signed-off-by: Soren Friis <soren.friis@intel.com>