atomisp2: remove 12x12 ispfilter setting for online video mode
BZ: 215519

Previously, this was added because video digital zoom needs it, but it looks
this is not the case for now. Video digital zoom without it works well. And
because there is a new video binary added for SAND which requires envelope
must be set to 0x0, thus chaning the envelope(ispfilter) to 0x0 by default
when DVS is turned off.

Change-Id: I66175311f7bcc7bf28bf211cf4f6ab7748919a8c
Signed-off-by: Bin Han <bin.b.han@intel.com>