atomisp2: add s3a stat buffer queue for HAL
Currently, ATOMISP_IOC_G_3A_STAT only supports camera HAL to get the latest
s3a statistics. But in some case, HAL may be blocked for some reason and
then previous statistics was overwritten by following one.

To avoid this situation, a queue with depth 8 was implemented just like the
metadata queue so that diver will not overwrite any s3a statistics until HAL
does not get it in 8 frame periods.

And this patch also removes the spinlock to protect the link lists because all
link list operations are happened within isp->mutex in hand, so it is not
necessary to do spinlock.

Change-Id: I937235cd72762c5f5ab34cf245fcf410739ec975
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-1980
Signed-off-by: Bin Han <bin.b.han@intel.com>
Signed-off-by: Benson.Yang <haipingx.yang@intel.com>
Reviewed-on: https://android.intel.com:443/260559