atomisp2: Revise crop region setting when YUV downscaling is needed.
If ISP input resolution is 10% bigger than required output resolution,
atomisp driver will try to use YUV downscaling.
But the logic of configuring crop region is not correct for this case.
When scaling is needed, the ISP input is first cropped to a size
which has the same aspect ratio as the required output size.
Then scaling is performed to make ISP output required resolution.

Change-Id: Iaf93371aa84eb27b264052f1b4b1779240ae0000
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-4314
Signed-off-by: Tianshu Qiu <tian.shu.qiu@intel.com>
Reviewed-on: https://android.intel.com:443/288187