imx: Add power off/on judgment  in flip setting
If the imx was powerd off, the flip setting function would request timeout
because of I2C transfers. It would lead to system start time too long and
other issues. So we should add a judgment in the flip setting functiong.
If the imx was powerd off, the flip setting function directly returned
error.

Change-Id: I8f2e7042be1a596efe39e79792af55a676a3b9be
Signed-off-by: Liu,Dalei <daleix.liu@intel.com>
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-776
Reviewed-on: https://android.intel.com:443/289231