atomisp2: Apply digital zoom as global settings immediately
There is a change that if HAL sets per-frame settings before
the last digital zoom setting is updated to css, the zoom
setting will be used as per-frame setting instead of global
setting. This is not expected since we need to treat zoom
setting as global setting.
This patch fixes this issue by applying digital zoom setting
immediately to css when it is set.

Change-Id: I888eb0f199bd8c28a60c7af32daa4883cf6205e1
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-776
Signed-off-by: Tianshu Qiu <tian.shu.qiu@intel.com>
Reviewed-on: https://android.intel.com:443/288509