atomisp2: Do not set sc_table as null if LSC is disabled
From the implementation of CSS, if sc_table is set to null
and passed to CSS. It means that the sc_table should not be
updated. This will not actually disable LSC. CSS just uses
the previous LSC table.
To disable LSC, driver needs to pass a valid sc_table with
enable=0 to CSS. CSS then can actually disable the LSC table.

Change-Id: Ica757881016a729b27ad79966cf1e794468fc973
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-797
Signed-off-by: Tianshu Qiu <tian.shu.qiu@intel.com>
Reviewed-on: https://android.intel.com:443/291027