atomisp2: poll for pipeline to stop after stopping stream
When the stream is stopped through calling ia_css_stream_stop(), the flag for SP to stop
the pipeline is set and returns immediately. It is non-blocking call. There could be a
race happening that PM turns off the power rails to I-UNIT or some voltage fluctuation
happens if we return the status of stream stopped too soon or is not coordinated. Then
driver is supposed to poll for pipeline to stop through ia_css_stream_has_stopped().

Change-Id: I28c62be5d4f2a0301cd3a1ebac658aa440056f21
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-4324
Signed-off-by: Shuguang Gong <shuguang.gong@intel.com>
Reviewed-on: https://android.intel.com:443/294306