atomisp2: Revert cropping region calculation fixes for MOFD
A regression is found on MOFD when the cropping region
calculation patches are applied.
The reason of the regression is due to strict FW limitation
protection after the patches are applied.
Since it is urgent for MOFD to pass CTS test cases, a temporary
workaround is added here.
The cropping region patches are "reverted" only for MOFD platform.
The patches are still required for IRDA and other projects.

Change-Id: I5c541e0f1a7a87fcdf8a56e187e1df3b04f2c23f
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-5610
Signed-off-by: Tianshu Qiu <tian.shu.qiu@intel.com>
Reviewed-on: https://android.intel.com:443/295986