atomisp2: [WORKAROUND]Re-trigger capture in ZSL mode when multiple buffers are queued
Due to the design of HALv3, sometimes in ZSL mode HAL needs to
capture multiple images within one streaming cycle.
But the capture number cannot be determined by HAL.
So HAL only sets the capture number to be 1 and queue multiple
buffers. Atomisp driver needs to check this case and re-trigger
CSS to do capture when new buffer is queued.

Change-Id: Iba193e86d49dc9a71a1da1d8a2c4aa0422044808
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-6341
Signed-off-by: Tianshu Qiu <tian.shu.qiu@intel.com>
Reviewed-on: https://android.intel.com:443/301556