atomisp2: Fix allocation of metadata list
If snapshots are taken during video the memory allocated by
vmap is leaked. This is because atomisp_reqbuf is called every
time snapshot is taken and new metadata buffers are incorrectly
allocated. To prevent memory leaking it should be also checked if
metadata_ready and metadata_in_css lists are empty or not and
perform allocation only if necessary. Fix it.

Change-Id: I7ba68e1fbcc7cb628636e9af6c32f5602a64d0a1
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-7959
Signed-off-by: Tommi Franttila <tommi.franttila@intel.com>
Reviewed-on: https://android.intel.com:443/305938