atomisp2: parameter not updated until atomisp_load_acc_extension
Because atomisp_load_acc_extension will destroy stream/pipes and
re-create them in order the extended binary can be loaded to pipe,
this will cause ISP parameter lost which was configured before that.
The patch just changes that before really starting the stream, ISP
parameters are stored and a flag is used to keep the update need,
and after atomisp_load_acc_extension and before stream starts,
check the update flag and update ISP parameters.

Change-Id: I02ffa8549fda73a5912f393678e0afdb52ff7c16
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-14196
Signed-off-by: Bin Han <bin.b.han@intel.com>
Signed-off-by: Tommi Franttila <tommi.franttila@intel.com>
Reviewed-on: https://android.intel.com:443/317393