imx132: add 1336x1096 resolution to still capture table
Add resolution 1336x1096 to the still capture mode table of the
imx132 driver. This is to work around the problem that CSS
is not cropping properly the image (the cropped window is at the
left edge of the original image, not centered as it should be).

This patch can be reverted once the cropping issue is fixed.

Change-Id: I83c415d7288ceec66c92a49712ed84ca6b48ae6a
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-14443
Signed-off-by: Tuukka Toivonen <tuukka.toivonen@intel.com>
Reviewed-on: https://android.intel.com:443/323102